website/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-appconf==1.0.4
django-classy-tags==2.0.0
django-cms==3.8.0
django-compressor==2.4
django-filer==2.0.2
django-formtools==2.2
django-jet==1.0.8
django-js-asset==1.2.2
django-mptt==0.11.0
django-multi-email-field==0.6.2
django-parler==2.2
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.7.1
html5lib==1.1
Pillow==8.1.0
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.5
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
Unidecode==1.1.2
webencodings==0.5.1